+++ src/main/java/org/mjulikelion/engnews/dto/ai/request/AnalyzeSentenceDto.java
@@ -1,4 +1,4 @@
-package org.mjulikelion.engnews.dto.ai;
+package org.mjulikelion.engnews.dto.ai.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class E2kTranslateDto {
-    private String news_content;
+public class AnalyzeSentenceDto {
+    private String news_sentence;
 }
