+++ src/main/java/org/mjulikelion/engnews/dto/ai/request/K2eTranslateDto.java
@@ -1,4 +1,4 @@
-package org.mjulikelion.engnews.dto.ai;
+package org.mjulikelion.engnews.dto.ai.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class TrySummarizeDto {
-    String news_content;
-    String message;
+public class K2eTranslateDto {
+    private String news_content;
 }
