+++ src/main/java/org/mjulikelion/engnews/dto/ai/request/TranslateDto.java
@@ -1,4 +1,4 @@
-package org.mjulikelion.engnews.dto.ai;
+package org.mjulikelion.engnews.dto.ai.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class TryTranslateDto {
-    String news_sentence;
-    String message;
+public class TranslateDto {
+    String news_content;
 }
