+++ src/main/java/org/mjulikelion/engnews/dto/ai/request/TrySummarizeDto.java
@@ -1,4 +1,4 @@
-package org.mjulikelion.engnews.dto.ai;
+package org.mjulikelion.engnews.dto.ai.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AnalyzeSentenceDto {
-    private String news_sentence;
+public class TrySummarizeDto {
+    String news_content;
+    String message;
 }
