+++ src/main/java/org/mjulikelion/engnews/dto/ai/request/TryTranslateDto.java
@@ -1,4 +1,4 @@
-package org.mjulikelion.engnews.dto.ai;
+package org.mjulikelion.engnews.dto.ai.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class K2eTranslateDto {
-    private String news_content;
+public class TryTranslateDto {
+    String news_sentence;
+    String message;
 }
