+++ src/main/java/org/mjulikelion/engnews/dto/ai/response/K2eResponseDto.java
@@ -1,4 +1,4 @@
-package org.mjulikelion.engnews.dto.ai;
+package org.mjulikelion.engnews.dto.ai.response;
 
 import lombok.Builder;
 import lombok.Getter;
